requirements.txt
django
django-filter
markdown
djangorestframework
djangorestframework-simplejwt
Faker
